Add files via upload
This script helps clean up data files so they can be easily put into SQL Server. It looks through all the folders and files you give it, and fixes any files it knows how to handle (like Excel, CSV, JSON, and text files). The main job of the script is to remove or change characters that can cause problems in SQL Server. These "bad" characters include things like quotation marks, special symbols, and punctuation. The script replaces these with underscores to make the data safe for SQL Server. It works on different types of files, cleaning them all in a similar way. After cleaning a file, it deletes the old one and makes a new, clean file with "out_" at the start of its name. This helps you know which files are ready to use. The script does all this automatically, which saves a lot of time and reduces mistakes when preparing data for SQL Server. It's especially helpful when you have lots of files or data to clean up.